% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSelect.R
\name{DESCENDselect}
\alias{DESCENDselect}
\title{DESCEND gene selection}
\usage{
DESCENDselect(counts, n.cores = 1, threshold = 3)
}
\arguments{
\item{counts}{the cell-by-gene expression counts. 
Note that DESCEND uses a Poisson model, so the count data should be provided (without normalization or log transformation).}

\item{n.cores}{the number of cores used for parallel computing. DESCEND can be slow so parallelization is highly recommended.}

\item{threshold}{the threshold for Gini index. Higer threshold leads to fewer selected genes.}
}
\value{
A list containing \describe{
  \item{select.genes}{the names of selected genes, ordered by decreasing scores.}
  \item{scores}{a score matrix containing the Gini scores of all genes.}
}
}
\description{
Select highly variable genes for clustering using DESCEND.
}
