% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOUPlineage.R
\name{get_connections}
\alias{get_connections}
\title{Helper function}
\usage{
get_connections(clus, forest, parent = NULL)
}
\arguments{
\item{forest}{the adjacency matrix}

\item{parent}{parent that will be excluded from the result}

\item{clust}{the cluster to start with}
}
\value{
The list of descendants
}
\description{
Extract all descendants starting from `clus`
}
