% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOUP.R
\name{getCenters}
\alias{getCenters}
\title{Get cluster centers}
\usage{
getCenters(expr, theta)
}
\arguments{
\item{expr}{a cell-by-gene expression matrix, either the raw counts or log-transformed expressions.}

\item{theta}{the cell-by-K membership matrix}
}
\value{
A K-by-gene matrix representing the centers of K clusters.
}
\description{
Compute the centers of clusters.
}
