% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOUPlineage.R
\name{getClusterLineages}
\alias{getClusterLineages}
\title{Get lineages from cluster centers}
\usage{
getClusterLineages(centers, end.clust = NULL, start.clust = NULL,
  dist.fun = NULL)
}
\arguments{
\item{centers}{K-by-p matrix of cluster centers}
}
\description{
Given cluster centers, construct a linearge using minimum-spanning tree.
A starting or ending cluster can be specified.
This code is adapted from the `getLineages` method in R package `slingshot`.
}
