% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOUPtime.R
\name{getClusterOrder}
\alias{getClusterOrder}
\title{Order clusters}
\usage{
getClusterOrder(cor.center, k.start = 1)
}
\arguments{
\item{cor.center}{The K-by-K correlation matrix among cluster centers}

\item{k.start}{The cluster to start with}
}
\value{
A list with length K that re-orders the clusters.
}
\description{
Re-order cluster labels, starting from \code{k.start}, 
and consecutively append the next one with the highest correlation in cluster centers.
}
