% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOUPlineage.R
\name{getLineageTime}
\alias{getLineageTime}
\title{Decide cell pseudotime}
\usage{
getLineageTime(lineages, membership)
}
\arguments{
\item{lineages}{A list of vectors, one per lineage, 
each containing the ordered cluster labels along the lineage}

\item{membership}{the n-by-K soft membership matrix}
}
\value{
An n-by-L pseudotime matrix for L lineages. 
    Each column contains the pseudotime of cells along one lineage, and NA if a cell does not belong to it. 
    Cells are assigned to different lineages, potentially with overlaps, according to their major clusters.
}
\description{
Compute the pseudotime of each cell along each lineage.
}
