% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{getMajorMatrix}
\alias{getMajorMatrix}
\title{Get the SOUP majority matrix}
\usage{
getMajorMatrix(memberships, Ks, cell.type, type = "hard",
  ref.lab = "Reference")
}
\description{
Get the SOUP majority matrix
}
