% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOUP.R
\name{getTheta}
\alias{getTheta}
\title{Get the membership matrix and centers}
\usage{
getTheta(expr, i.pure, pure.cluster, G)
}
\arguments{
\item{expr}{a cell-by-gene expression matrix, either the raw counts or log-transformed expressions.}

\item{i.pure}{the indices of the pure cells..}

\item{pure.cluster}{a vector containing {1, ..., K}, indicating the cluster labels of pure cells.}

\item{G}{the cell-by-nPC matrix obtained from SVD of the similarity matrix.}
}
\value{
A list containing  \describe{
  \item{membership}{the cell-by-K membership matrix.}
  \item{centers}{the K-by-gene membership matrix.}
}
}
\description{
Compute the membership matrix and cluster centers based on pure cells.
}
