% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOUPtime.R
\name{getTimeline}
\alias{getTimeline}
\title{Estimate Developmental Timeline}
\usage{
getTimeline(membership, centers, k.start = NULL, k.end = NULL)
}
\arguments{
\item{membership}{An n.cell-by-K soft membership matrix}

\item{centers}{A K-by-n.gene cluster center matrix}

\item{k.start}{Which cluster is the starting point? 
One, and only one of \code{k.start} and \code{k.end} must be specified.}

\item{k.end}{Which cluster is the ending point?
One, and only one of \code{k.start} and \code{k.end} must be specified.}
}
\value{
A vector of timepoints for all cells.
}
\description{
Estimate the developmental timeline for cells from SOUP soft membership.
}
