% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{heatmapKseq}
\alias{heatmapKseq}
\title{Visualize SOUP hard assignments.}
\usage{
heatmapKseq(memberships, Ks, cell.type)
}
\arguments{
\item{memberships}{A list of membership matrices of different K}

\item{Ks}{A list of different Ks}

\item{cell.type}{The golden standard cell types for reference.}
}
\value{
A ggplot object.
}
\description{
Visualize SOUP hard assignments.
}
