% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{plotGeneTimeline}
\alias{plotGeneTimeline}
\title{Plot gene expression along trajectory}
\usage{
plotGeneTimeline(expr, marker.gene, timeline, x.title = "SOUP trajectory",
  y.title = "Expression", title = "", font.size = 10)
}
\arguments{
\item{expr}{The cell-by-gene expression matrix to be visualized}

\item{marker.gene}{The gene to be visualized; must be one of the columns of \code{expr}}

\item{timeline}{Timepoints for all cells, a vector}

\item{x.title}{(optional) Label of x-axis}

\item{y.title}{(optional) Label of y-axis}

\item{title}{(optional) Plot title}
}
\value{
A ggplot object
}
\description{
Plot gene expression along trajectory
}
