% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot_lineage.R
\name{plotLineageCurves}
\alias{plotLineageCurves}
\title{Plot lineages or curves}
\usage{
plotLineageCurves(x, centers = NULL, dims = c(1:2), add = FALSE,
  labels = NULL, lwd = 2, col = 1, lab.cex = 2, cex = 2,
  pos = 2, offset = 0.5, type = "lineages")
}
\description{
Plot lineages or curves
}
