% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{plotMultipleGeneTimeline}
\alias{plotMultipleGeneTimeline}
\title{Plot multiple gene expressions along trajectory}
\usage{
plotMultipleGeneTimeline(expr, genelist, timeline,
  x.title = "SOUP trajectory", y.title = "Expression", nrow = NULL,
  ncol = NULL)
}
\arguments{
\item{expr}{The cell-by-gene expression matrix to be visualized}

\item{genelist}{A list of genes to be visualized; must be among the columns of \code{expr}}

\item{timeline}{Timepoints for all cells, a vector}

\item{x.title}{(optional) Label of x-axis}

\item{y.title}{(optional) Label of y-axis}

\item{nrow}{(optional) number of rows in the plot grid.}

\item{ncol}{(optional) number of cols in the plot grid.}
}
\value{
A ggplot object
}
\description{
Expression levels of multiple genes along trajectory, arranged on a grid of plots.
}
