% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvSOUP.R
\name{predictTheta}
\alias{predictTheta}
\title{Predict the membership for new data points}
\usage{
predictTheta(new.expr, t.centers)
}
\arguments{
\item{new.expr}{cell-by-gene expression matrix}

\item{t.centers}{transposed center matrix, n.gene-by-K}
}
\value{
The predicted membership matrix.
}
\description{
Predict the membership for new data points
}
