% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOUP.R
\name{projMembership}
\alias{projMembership}
\title{Clean up membership matrix}
\usage{
projMembership(theta)
}
\arguments{
\item{theta}{The estimated raw theta}
}
\value{
The cleaned-up membership matrix.
}
\description{
Clean up membership matrix
}
