% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOUP.R
\name{scaleRowSums}
\alias{scaleRowSums}
\title{Scale matrix by row sums}
\usage{
scaleRowSums(x)
}
\arguments{
\item{x}{a matrix}
}
\value{
the scaled matrix.
}
\description{
Scale matrix such that each row sums to 1.
}
