% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSelect.R
\name{selectGenes}
\alias{selectGenes}
\title{Select highly variable genes}
\usage{
selectGenes(expr, type = "count", SPCA = TRUE, DESCEND = TRUE,
  n.cores = 1, threshold = 3, sumabs = 0.05, nPC = 3)
}
\arguments{
\item{expr}{a cell-by-gene expression matrix, either the raw counts or log-transformed expressions.}

\item{type}{"count" if \code{expr} contains the raw counts (default), 
or "log" if \code{expr} has been normalized and log-transformed.}

\item{SPCA}{boolean, whether to use SPCA or not.}

\item{DESCEND}{boolean, whether to use DESCEND or not.}

\item{n.cores}{the number of cores used for parallel computing of DESCEND. 
DESCEND can be slow so parallelization is highly recommended.}

\item{threshold}{the threshold for Gini index of DESCEND. Higer threshold leads to fewer selected genes.}

\item{sumabs}{a measurement of sparsity of genes in SPCA, between \code{1/sqrt(n.gene)} and 1.
Smaller values result in sparser results, hence fewer selected genes.}

\item{nPC}{the number of sparse singular vectors to use in SPCA.}
}
\value{
A list of gene names that are selected.
}
\description{
Select highly variable genes for SOUP clustering, using SPCA and/or DESCEND.
}
