% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOUP.R
\name{SOUP}
\alias{SOUP}
\title{SOUP}
\usage{
SOUP(expr, Ks = 3, type = "count", i.pure = NULL, ext.prop = NULL,
  pure.prop = 0.5, nPC = NULL)
}
\arguments{
\item{expr}{a cell-by-gene expression matrix, either the raw counts or log-transformed expressions.}

\item{Ks}{number of clusters, can be a single integer or a list of integers.}

\item{type}{"count" (default) if \code{expr} contains the raw counts, or "log" if \code{expr} has been normalized and log-transformed.}

\item{i.pure}{(optional) the indices of the pure cells. By default is \code{NULL}, and SOUP will infer the pure list.
If the list is already known (for example, from previous runs), then providing it will reduce the computation time.}

\item{ext.prop}{(optional) the proportion of extreme neighbors for each cell, such that \code{ext.prop*n.cells} is roughly the number of pure cells \emph{per cluster}. 
By default, \code{ext.prop=0.1} for less than 1,000 cells, and \code{ext.prop=0.05} for larger datasets.}

\item{pure.prop}{(optional) the expected proportion of pure cells in the data. By default \code{pure.prop=0.5}.}

\item{nPC}{(optional) the number of principal components to be used, which is at least \code{K} (default). 
Sometimes it can be helpful to use a slightly larger \code{nPC}.}
}
\value{
A list containing  \describe{
  \item{Ks}{the list of \code{K}, number of clusters, used by SOUP.}
  \item{memberships}{a list of cell-by-K membership matrices, one per \code{K} in \code{Ks}.}
  \item{centers}{a list of K-by-gene cluster center matrices, one per \code{K} in \code{Ks}.}
  \item{purity}{a vector containing the purity scores of all cells.}
  \item{i.pure}{the indices of pure cells with the highest purity scores.}
}
}
\description{
A semi-soft clustering algorithm for single cells.
}
\examples{
select.genes = zeisel$select.genes
counts = zeisel$counts[, colnames(counts) \%in\% select.genes]
soup.out = SOUP(counts, Ks=7, type="count")

}
