% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseSOUP.R
\name{sparseSOUP}
\alias{sparseSOUP}
\title{Sparse SOUP}
\usage{
sparseSOUP(expr, Ks = 3, type = "log", delta = seq(0.1, 2.5, 0.1),
  i.pure = NULL, ext.prop = NULL, pure.prop = 0.5, nPC = NULL,
  nstart = 50, nCV = 50, verbose = FALSE)
}
\arguments{
\item{expr}{a cell-by-gene expression matrix, either the raw counts or log-transformed expressions.}

\item{Ks}{number of clusters, can be a single integer or a list of integers.}

\item{type}{"log" if \code{expr} has been normalized and log-transformed (default),
or "count" if \code{expr} contains the raw counts.
It is recommended to use the log scale, which usually gives better results in practice.}

\item{i.pure}{(optional) the indices of the pure cells. By default is \code{NULL}, and SOUP will infer the pure list.
If the list is already known (for example, from previous runs), then providing it will reduce the computation time.}

\item{ext.prop}{(optional) the proportion of extreme neighbors for each cell, such that \code{ext.prop*n.cells} is roughly the number of pure cells \emph{per cluster}. 
By default, \code{ext.prop=0.1} for less than 1,000 cells, and \code{ext.prop=0.05} for larger datasets.}

\item{pure.prop}{(optional) the expected proportion of pure cells in the data. By default \code{pure.prop=0.5}.}

\item{nPC}{(optional) the number of principal components to be used, which is at least \code{K} (default). 
Sometimes it can be helpful to use a slightly larger \code{nPC}.}
}
\value{
A list containing  \describe{
  \item{Ks}{the list of \code{K}, number of clusters, used by SOUP.}
  \item{memberships}{a list of cell-by-K membership matrices, one per \code{K} in \code{Ks}.}
  \item{centers}{a list of K-by-gene cluster center matrices, one per \code{K} in \code{Ks}.}
  \item{purity}{a vector containing the purity scores of all cells.}
  \item{i.pure}{the indices of pure cells with the highest purity scores.}
}
}
\description{
A sparse semi-soft clustering algorithm for single cells.
}
