% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOUPcv.R
\name{cv.error.SOUP}
\alias{cv.error.SOUP}
\title{Compute cross validation errors}
\usage{
cv.error.SOUP(expr, type = "log", Ks = c(2:10), nfold = 10, seed = NULL,
  mc.cores = 1)
}
\arguments{
\item{expr}{A cell-by-gene expression matrix, either the raw counts or log-transformed expressions.}

\item{type}{"log" if \code{expr} has been normalized and log-transformed (default),
or "count" (default) \code{expr} contains the raw counts.}

\item{Ks}{A sequence of cluster numbers}

\item{nfold}{(optional) Number of folds, default is 10}

\item{seed}{(optional) random seed, default is NULL}

\item{mc.cores}{(optional) Number of cores for parallelization, default is 1 without parallelization}
}
\description{
Compute cross validation errors
}
