% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvSOUP.R
\name{cvSOUP}
\alias{cvSOUP}
\title{Cross validation for SOUP}
\usage{
cvSOUP(expr, type = "log", Ks = c(2:10), nfold = 10, nCV = 10,
  mc.cores = 1, seeds = NULL, verbose = TRUE)
}
\arguments{
\item{expr}{A cell-by-gene expression matrix, either the raw counts or log-transformed expressions.}

\item{type}{"log" if \code{expr} has been normalized and log-transformed (default),
or "count" (default) \code{expr} contains the raw counts.}

\item{Ks}{A sequence of cluster numbers}

\item{nfold}{(optional) Number of folds, default is 10}

\item{nCV}{(optional) Number of repetitions, default is 10}

\item{mc.cores}{(optional) Number of cores for parallelization, default is 1 without parallelization}

\item{seeds}{(optional) A list of seeds to be used, with length nCV, default is NULL}

\item{verbose}{(optional) Whether to print progress, default is TRUE}
}
\description{
Cross validation for SOUP
}
