% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOUP.R
\name{findPure}
\alias{findPure}
\title{Find pure cells}
\usage{
findPure(A, ext.prop = NULL, pure.prop = 0.5)
}
\arguments{
\item{A}{cell-by-cell similarity matrix.}

\item{ext.prop}{(optional) the proportion of extreme neighbors for each cell, such that \code{ext.prop*n.cells} is roughly the number of pure cells \emph{per cluster}. 
By default, \code{ext.prop=0.1} for less than 1,000 cells, and \code{ext.prop=0.05} for larger datasets.}

\item{pure.prop}{(optional) the proportion of pure cells in the data. By default \code{pure.prop=0.5}.}
}
\value{
A list containing \describe{
  \item{i.pure}{indices for the top \code{pure.prop*n.cells} pure cells.}
  \item{purity}{purity scores of all cells.}
}
}
\description{
Find the list of pure cells with the highest purity scores.
}
