% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOUP.R
\name{getSimilarity}
\alias{getSimilarity}
\title{Compute similarity matrix}
\usage{
getSimilarity(expr, type = "count")
}
\arguments{
\item{expr}{a cell-by-gene expression matrix, either the raw counts or log-transformed expressions.}

\item{type}{"count" (default) if \code{expr} contains the raw counts, or "log" if \code{expr} has been normalized and log-transformed.}

\item{method}{the method to measure similarity between cells, one of \{"cov", "cor", "inner\},}
}
\value{
The cell-cell similarity matrix.
}
\description{
Compute the cell-cell similarity matrix.
}
